rfctr: move multipage_sections to by-title
This is by-title-specific, but more importantly it controls the choice
of boundary-predicates rather than directly being an option. Move it to
the `_ByTitleChunkingOptions`.
